+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Changelog for the gruf gem. This includes internal history before the gem was ma
 
 ### Pending release
 
+- [#163] Add `context` hash attribute to `Gruf::Controllers::Request` to allow interceptors to pass information down
+  to a gruf controller
+
 ### 2.15.1
 
 - Fix issue where GRPC_SERVER_POOL_KEEP_ALIVE and GRPC_SERVER_POLL_PERIOD when set via ENV are not cast to int